Close rethink db connections.

+++ src/main/java/fund/cyber/node/model/MigrationLog.java
@@ -1,30 +1,30 @@
 package fund.cyber.node.model;
 
 public class MigrationLog {
-    private int id;
-    private int hash;
+    private long id;
+    private long hash;
 
     public MigrationLog() {
     }
 
-    public MigrationLog(int id, int hash) {
+    public MigrationLog(long id, long hash) {
         this.id = id;
         this.hash = hash;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getHash() {
+    public long getHash() {
         return hash;
     }
 
-    public void setHash(int hash) {
+    public void setHash(long hash) {
         this.hash = hash;
     }
 }
